airflow/requirements/requirements.txt
google.cloud
apache-airflow-providers-google
apache-airflow[gcp]
apache-airflow[google]
datetime
pandas
flask
flask_admin